refactor: move the methods namespace to extensions/methods namespace
